add avoqado api

+++ app/src/main/java/com/avoqado/pos/screens/tableDetail/model/TableDetail.kt
@@ -6,7 +6,7 @@ data class TableDetail(
     val tableId: String = "",
     val name: String = "",
     val products: List<Product> = emptyList(),
-    val totalAmount: Float = 0.0f
+    val totalAmount: Double = 0.0
 ) {
     val formattedTotalPrice : String
         get() = StringUtils.toStringThousandAmount(StringUtils.notFormatAmount(totalAmount.toString()))
